[FIX] missing awareness of complex queries
The _fix_condition() hack was meant as a
usability thing to easy explaining the
query API to users without going into
details of SQL::Abstract due to the
visibility of contained JOINs.

However, the way to 'fix' that is very fragile.

Here we fix one more obvious case when
using 'like' searches but we will also
provide a complete optional disabling of that
in the next patches.