[FEAT] allow pure SQL::Abstract queries in query api
The _fix_condition() hack was meant as a usability
thing to easy explaining the query API to users
without going into details of SQL::Abstract due
to the visibility of contained JOINs.

The implementation, however, is fragile as it
tries to patch the queries with regexes which
obviously is very lame.

To skip that mechanism we now allow a parameter
in the API 'puresqlabstract=1' and pass it
through to the Tapper DPath handling.