[FIX] die hard with a vengance
When an error occurs while processing a given template an error message
was generated and printed. Printing error messages is only allowed in
frontend functions (mainly CLI). Backend functions must never print.

With this patch we now die in case of an error. This is the same that
happens for template file processing and is what the calling Cmd
function already expects.