sorted keys in virtual function 'Dumper'
Dumper() is available in Mason and TT templates
and provides a Data::Dumper output of data.

We now by default sort the keys. This takes the
burden from the templates to make the output more
readable.

The Dumper is usually only used for template development
and debugging. The production code would usually use
json or yaml anyway.