TT - testrundata_nohost() to not JOIN host table
The function testrundata() contains host data,
so you can filter on it.

However, to get testruns that are not yet
scheduled and therefore do not yet have a
host assigned, we need to not JOIN the host
table.

In particular you would never get testruns
in states

  - prepare
  - schedule

with the original function.

For backwards compatibility we provide a
dedicated function

  testrundata_nohost()

which passes a newly introduced flag to the
backend functions that handle this.

With that new function you also get those testruns.